third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-drawImage-video-expected.txt

Ensure correct behavior of drawImage with ImageBitmaps from video elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[3] is 255
PASS d[3] is 255
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS jsWrapperClass(bitmap) is 'ImageBitmap'
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS d[0] is 0
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
